reformat 6: electric boogalix
