+++ Assets/Scripts/Collectibles/Collectible.cs
@@ -1,0 +1,16 @@
+using UnityEngine;
+
+namespace Collectibles
+{
+    public abstract class Collectible : MonoBehaviour
+    {
+        protected virtual void OnTriggerEnter2D(Collider2D collision)
+        {
+            if (!collision.gameObject.CompareTag("Player")) return;
+            Pickup(collision.gameObject);
+            Destroy(this.gameObject);
+        }
+
+        protected abstract void Pickup(GameObject player);
+    }
+}
