+++ Assets/Scripts/Collectibles/RandomAmmoPickup.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+namespace Collectibles
+{
+    public class RandomAmmoPickup : AmmoPickup
+    {
+        [SerializeField] [Range(1, 50)] private int minAmt;
+        [SerializeField] [Range(1, 50)] private int maxAmt;
+
+        private void Awake()
+        {
+            dropAmount = Random.Range(minAmt, maxAmt);
+        }
+    }
+}
